requirements.txt
pip>=21.1.2
configparser>=5.0.2
wheel>=0.34.2
cryptography>=2.8
lxml>=4.5.0
tornado>=6.1
keyring>=18.0.1
six>=1.14.0
fire>=0.4.0
termcolor>=1.1.0
colorama>=0.4.3
pywebcopy>=6.3.0
requests>=2.22.0
w3lib>=1.22.0
pytube>=10.8.2
pyquery>=1.4.3
cssselect>=1.1.0
bs4>=0.0.1
beautifulsoup4>=4.8.2
parse>=1.19.0
setuptools>=57.0.0
wget>=3.2

youtube-dl
python-telegram-bot
docker